pyproject.toml
[tool.poetry]
name = "portrait_search"
version = "0.0.1"
description = "OpenAI use case for natural language image search"
authors = ["Anatolii Stehnii <tsdaemon@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
aiohttp = "^3.9.1"
aiofiles = "^23.2.1"
tqdm = "^4.66.1"
rarfile = "^4.1"
pyyaml = "^6.0.1"
py7zr = "^0.20.8"
motor = "^3.3.2"
imagehash = "^4.3.1"
dependency-injector = "^4.41.0"
openai = "^1.6.1"
backoff = "^2.2.1"
loguru = "^0.7.2"
sentence-transformers = "^2.2.2"
langchain = "^0.0.352"
instructorembedding = "^1.0.1"
pydantic-mongo = "^2.0.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
types-tqdm = "^4.66.0.5"
types-aiofiles = "^23.2.0.0"
types-pyyaml = "^6.0.12.12"
ruff = "^0.1.9"
types-pillow = "^10.1.0.2"
motor-types = "^1.0.0b4"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
aioresponses = "^0.7.6"
pytest-asyncio = "^0.23.2"
mock = "^5.1.0"
ipython = "^8.19.0"
types-mock = "^5.1.0.3"

[tool.poetry.group.experiment.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 120

select = ["E", "F", "I"]
ignore = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".data",
    ".mypy_cache",
    "experiments",
]
per-file-ignores = {}

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.11.
target-version = "py311"
